requirements.txt
openai==1.65.4
requests==2.32.3
pydantic==2.10.6
graphviz==0.20.3
IPython==9.0.1
pytz==2025.1
llama-index==0.12.23
llama-index-llms-openai-like==0.3.4
langchain==0.3.20
langchain-openai==0.3.8
langgraph==0.3.5
langgraph-cli[inmem]==0.1.75
debugpy==1.8.13
autogen-agentchat==0.4.8.2
autogen-ext[openai]==0.4.8.2
semantic-kernel==1.23.1
pyperclip==1.9.0
python-dotenv==1.0.1